+++ src/extension.ts
@@ -11,6 +11,7 @@ import { WebSocketServer } from "./hardis-websocket-server";
 import { Logger } from "./logger";
 import { getWorkspaceRoot, preLoadCache } from "./utils";
 import { WelcomePanel } from "./webviews/welcome";
+import { HardisColors } from "./hardis-colors";
 
 let refreshInterval: any = null;
 
@@ -69,6 +70,10 @@ export function activate(context: vscode.ExtensionContext) {
   );
   context.subscriptions.push(...commands.disposables);
 
+  const hardisColors = new HardisColors();
+  context.subscriptions.push(hardisColors);
+  hardisColors.init();
+
   // Initialize Hardis Debugger commands
   const hardisDebugger = new HardisDebugger();
   context.subscriptions.push(...hardisDebugger.disposables);
@@ -109,7 +114,14 @@ export function activate(context: vscode.ExtensionContext) {
 
   // Catch event configuration changes
   vscode.workspace.onDidChangeConfiguration((event) => {
-    manageWebSocketServer();
+    if (event.affectsConfiguration("vsCodeSfdxHardis")) {
+      if (event.affectsConfiguration("vsCodeSfdxHardis.userInput")) {
+        manageWebSocketServer();
+      }
+      if (event.affectsConfiguration("vsCodeSfdxHardis.disableVsCodeColors")) {
+        hardisColors.init();
+      }
+    }
   });
 
   // Refresh commands if a sfdx-Project.json has been added