+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## [Unreleased]
 
+## [1.26.0] 2023-04-18
+
+- Manage VsCode Theme colors depending on selected org
+  - Automated
+    - Production: red
+    - Sandbox of major org (where you are not supposed to deploy ^^): orange
+    - Sandbox or scratch org for dev/config: green
+    - Other (Dev Org, Trial org...): blue
+  - Notes
+    - If you don't want this feature, use `vsCodeSfdxHardis.disableVsCodeColors` VsCode setting
+    - You can force the association between an org and a color using shortcut button in status panel
+- Rename master branch into main
+
 ## [1.25.1] 2023-03-13
 
 - Display hammer-wrench icon in dependencies section when a plugin is locally developed